% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple-bnab-BH-titer.R
\name{calc_2bnab_BHtiter}
\alias{calc_2bnab_BHtiter}
\title{Theoretical calculation for experimental (pooled sera) neutralization titer
under BH interaction for 2 antibodies (closed form solution)}
\usage{
calc_2bnab_BHtiter(pt50_1, pt50_2, titer_target = 0.5, NaN_as_NA = F)
}
\arguments{
\item{titer_target}{goal combination titer endpoint: 0.5 or 50 = PT50, 0.8 or 80 = PT80. Pct equivalence fails for tiny titer levels (<PT1): if you want PT<1, must input < 0.01}

\item{NaN_as_NA}{arbitrarily large titers returned as missing

This function is vectorized for use over PK}

\item{PT50_1}{ratio of concentration to IC50 for first Ab}

\item{PT50_2}{ratio of concentration to IC50 for second Ab}
}
\value{

}
\description{
Theoretical calculation for experimental (pooled sera) neutralization titer
under BH interaction for 2 antibodies (closed form solution)
}
