% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple-bnab-BH-titer.R
\name{calc_3bnab_BHtiter}
\alias{calc_3bnab_BHtiter}
\title{Theoretical calculation for experimental (pooled sera) neutralization titer
under BH interaction for 3 antibodies}
\usage{
calc_3bnab_BHtiter(
  pt50_1,
  pt50_2,
  pt50_3,
  titer_target = 50,
  min_titer_return = 0.01,
  NaN_as_NA = F
)
}
\arguments{
\item{pt50_1}{ratio of concentration to IC50 for first Ab}

\item{pt50_2}{ratio of concentration to IC50 for second Ab}

\item{pt50_3}{ratio of concentration to IC50 for third Ab}

\item{titer_target}{goal combination titer endpoint: 0.5 or 50 = PT50, 0.8 or 80 = PT80. Pct equivalence fails for tiny titer levels (<PT1): if you want PT<1, must input < 0.01}

\item{min_titer_return}{smallest triple BH titer returned (default 0.01)}

\item{NaN_as_NA}{arbitrarily large titers returned as missing

This function is vectorized for use over PK}
}
\value{

}
\description{
Theoretical calculation for experimental (pooled sera) neutralization titer
under BH interaction for 3 antibodies
}
