% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neut-funs.R
\name{convert_neut_iip}
\alias{convert_neut_iip}
\title{Calculate IIP given neutralization}
\usage{
convert_neut_iip(neut)
}
\arguments{
\item{neut}{neutralization between 0 and 1}
}
\value{
scalar or numerical vector
}
\description{
Calculate IIP given neutralization
}
