% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neut-funs.R
\name{convert_neut_pt50}
\alias{convert_neut_pt50}
\title{Calculate predicted titer 50 (pt50) given neutralization}
\usage{
convert_neut_pt50(neut, hill = 1)
}
\arguments{
\item{neut}{neutralization between 0 and 1}
}
\value{
scalar or numerical vector
}
\description{
Calculate predicted titer 50 (pt50) given neutralization
}
