% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neut-funs.R
\name{convert_pt50_neut}
\alias{convert_pt50_neut}
\title{Calculate neutralization given predicted titer 50}
\usage{
convert_pt50_neut(pt50, hill = 1)
}
\arguments{
\item{pt50}{positive value, ratio of concentration of IC80}
}
\value{
scalar or numerical vector
}
\description{
Calculate neutralization given predicted titer 50
}
