% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neut-funs.R
\name{convert_pt80_iip}
\alias{convert_pt80_iip}
\title{Calculate IIP given PT80 (for input into AMP PE curve)}
\usage{
convert_pt80_iip(pt, hill = 1)
}
\arguments{
\item{pt80}{positive value, ratio of concentration of IC80}
}
\value{
scalar or numerical vector
#' Calculate general predicted titer given neutralization
#' @param neut neutralization between 0 and 1
#' @param titer_level between 1 and 100
#' @return scalar or numerical vector
convert_neut_pt = function(neut, titer_level = 50, hill = 1) {
  4^(-1/hill*xx) * neut/(1 - neut)
}

#' Calculate neutralization given generalized titer
#' @param pt positive value, ratio of concentration of given IC
#' @param titer_input level for PT input (ex., 50 -> PT50)
#' @param titer_target level for PT output (ex., 80 -> PT80)
#' @param .titer_as_pct default T, assumed input is percent, converts values < 1
#' @return scalar or numerical vector
convert_titer_base = function(pt,
                              titer_input = 50, titer_target = 80,
                              hill = 1, .titer_as_pct = T) {
  stopifnot(titer_input > 0 & titer_input < 100)
  if(.titer_as_pct & titer_input < 1) titer_input = titer_input * 100
  stopifnot(titer_target > 0 & titer_target < 100)
  if(.titer_as_pct & titer_target < 1) titer_target = titer_target * 100

}

convert_pt50_pt80 = function(pt50, hill = 1) convert_titer_base(pt50, 50, 80, hill = hill)
convert_pt80_pt50 = function(pt80, hill = 1) convert_titer_base(pt80, 80, 50, hill = hill)
#' Calculate neutralization given generalized titer
#' @param pt positive value, ratio of concentration of given IC
#' @param titer_level between 1 and 100
#' @param .titer_as_pct default T, assumed input is percent, converts values < 1
#' @return scalar or numerical vector
convert_pt_neut = function(pt, titer_level = 50, hill = 1, .titer_as_pct = T) {
  stopifnot(titer_level > 0 & titer_level < 100)
  if(.titer_as_pct & titer_level < 1) titer_level = titer_level * 100

  1 - 1 / (1 + (pt * xxx ^ (1 / hill)) ^ hill)
}
}
\description{
Calculate IIP given PT80 (for input into AMP PE curve)
}
