% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neut-funs.R
\name{convert_pt80_neut}
\alias{convert_pt80_neut}
\title{Calculate neutralization given predicted titer 80}
\usage{
convert_pt80_neut(pt80, hill = 1)
}
\arguments{
\item{pt80}{positive value, ratio of concentration of IC80}
}
\value{
scalar or numerical vector
}
\description{
Calculate neutralization given predicted titer 80
}
